requirements.txt
# Data source
akshare>=1.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Data validation
pydantic>=2.0.0

# Visualization (Plotly for interactive HTML charts)
plotly>=5.18.0
kaleido>=0.2.0  # For static image export if needed

# CLI
click>=8.1.0
typer>=0.9.0

# Configuration
python-dotenv>=1.0.0

# MCP Server
mcp>=0.9.0

# Caching
diskcache>=5.6.0

# Data Export
openpyxl>=3.0.0  # Excel export support
